+++ src/rpc/util.cpp
@@ -777,3 +777,31 @@ EvalDescriptorStringOrObject(const UniValue &scanobject,
     }
     return ret;
 }
+
+UniValue GetServicesNames(ServiceFlags services) {
+    UniValue servicesNames(UniValue::VARR);
+
+    if (services & NODE_NETWORK) {
+        servicesNames.push_back("NETWORK");
+    }
+    if (services & NODE_GETUTXO) {
+        servicesNames.push_back("GETUTXO");
+    }
+    if (services & NODE_BLOOM) {
+        servicesNames.push_back("BLOOM");
+    }
+    if (services & NODE_XTHIN) {
+        servicesNames.push_back("XTHIN");
+    }
+    if (services & NODE_BITCOIN_CASH) {
+        servicesNames.push_back("BITCOIN_CASH");
+    }
+    if (services & NODE_NETWORK_LIMITED) {
+        servicesNames.push_back("NETWORK_LIMITED");
+    }
+    if (services & NODE_AVALANCHE) {
+        servicesNames.push_back("AVALANCHE");
+    }
+
+    return servicesNames;
+}